Add explicit cast for 64-to-32 bit conversion
TODO: use 64 bit mr sizes in capnp

+++ hydra/node.cpp
@@ -103,7 +103,7 @@ void node::recv(const request_t &request, const qp_t &qp) {
     info([&](const auto &info) {
       auto mr = msg.initInit().initInfo();
       mr.setAddr(reinterpret_cast<uintptr_t>(info.second->addr));
-      mr.setSize(info.second->length);
+      mr.setSize(static_cast<uint32_t>(info.second->length));
       mr.setRkey(info.second->rkey);
     });
     reply(qp, response);
